requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
selenium==4.15.2
webdriver-manager==4.0.1
Werkzeug==2.3.6
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1